refactor: 패키지 `mapper`, `validator` 분리
