+++ backend/src/main/java/reviewme/review/service/ReviewRegisterService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import reviewme.review.domain.Review;
 import reviewme.review.repository.ReviewRepository;
 import reviewme.review.service.dto.request.ReviewRegisterRequest;
-import reviewme.review.service.module.ReviewMapper;
-import reviewme.review.service.module.ReviewValidator;
+import reviewme.review.service.mapper.ReviewMapper;
+import reviewme.review.service.validator.ReviewValidator;
 
 @Service
 @RequiredArgsConstructor
