+++ backend/src/main/java/reviewme/review/service/validator/CheckBoxAnswerValidator.java
@@ -1,4 +1,4 @@
-package reviewme.review.service.module;
+package reviewme.review.service.validator;
 
 import java.util.HashSet;
 import java.util.List;
